clippy: fix Linux-only -D warnings in hi-tools sandbox and hf
The CI clippy step (on Linux) flagged three lints invisible on macOS:
- sandbox.rs is_enforced: needless return in the linux cfg block
- sandbox.rs wrap_program: tail fallthrough unreachable on linux (all cfg
  branches return) — kept the safe default, scoped allow to linux
- hf.rs available_space_bytes: f_bavail/f_frsize as u64 is a same-type cast on
  64-bit unix — allow unnecessary_cast since field width varies by target

Verified locally: exact CI clippy command exit 0, fmt --check clean, sandbox
and hf package tests pass.

+++ crates/hi-tools/src/hf.rs
@@ -980,7 +980,11 @@ pub fn available_space_bytes(path: &Path) -> Option<u64> {
         return None;
     }
     let stats = unsafe { stats.assume_init() };
-    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize))
+    // `f_bavail`/`f_frsize` widths differ across unix targets (u32 on some,
+    // u64 on Linux/macOS), so normalize through u64; where the field is
+    // already u64 the cast is a no-op, which clippy would otherwise flag.
+    #[allow(clippy::unnecessary_cast)]
+    Some((stats.f_bavail as u64).saturating_mul(stats.f_frsize as u64))
 }
 
 #[cfg(not(unix))]
